Support String Tensor in TFNet/TFTrainingHelper2 (#1803)
* tfnet support string

* fix style
