+++ zoo/src/test/resources/tfnet_string/graph_meta.json
@@ -1,0 +1,1 @@
+{"input_names": ["Placeholder:0"], "output_names": ["StringToNumber:0"]}
